testing/web-platform/mozilla/tests/service-workers/service-worker/clients-get.https.html

<!DOCTYPE html>
<title>Service Worker: Clients.get</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/get-host-info.sub.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
var host_info = get_host_info();

var scope = 'resources/blank.html?clients-get';
var t = async_test('Test Clients.get()');
var clientIds = [];
t.step(function() {
    service_worker_unregister_and_register(
        t, 'resources/clients-get-worker.js', scope)
      .then(function(registration) {
          return wait_for_state(t, registration.installing, 'activated');
        })
      .then(function() {
          return with_iframe(scope + '#1');
        })
      .then(function(frame1) {
          clientIds.push(frame1.contentDocument.body.textContent);
          frame1.focus();
          return with_iframe(scope + '#2');
        })
      .then(function(frame2) {
          clientIds.push(frame2.contentDocument.body.textContent);
          var channel = new MessageChannel();
          channel.port1.onmessage = t.step_func(on_message);
          frame2.contentWindow.navigator.serviceWorker.controller.postMessage(
              {port:channel.port2, clientIds:clientIds,
               message: 'get_client_ids'}, [channel.port2]);
        })
      .catch(unreached_rejection(t));
  });

var expected = [
    /* visibilityState, focused, url, frameType */
    ['visible', true, new URL(scope + '#1', location).toString(), 'nested'],
    ['visible', false, new URL(scope + '#2', location).toString(), 'nested'],
    undefined
];

function on_message(e) {
  assert_equals(e.data.length, 3);
  assert_array_equals(e.data[0], expected[0]);
  assert_array_equals(e.data[1], expected[1]);
  assert_equals(e.data[2], expected[2]);
  service_worker_unregister_and_done(t, scope);
}
</script>
